requirements.txt
numpy==1.24.3
pandas==2.0.3
pyxdf==1.0.2
pyyaml==6.0
scipy==1.11.1
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1
mne>=1.3.0
neurokit2>=0.2.0
xgboost>=1.7.0